usnic: Move extension code into separate file.
