+++ prov/usnic/Makefile.include
@@ -93,7 +93,8 @@ _usnic_files = \
 	prov/usnic/src/usdf_timer.c \
 	prov/usnic/src/usdf_timer.h \
 	prov/usnic/src/usdf_poll.c \
-	prov/usnic/src/usdf_poll.h
+	prov/usnic/src/usdf_poll.h \
+	prov/usnic/src/usdf_ext.c
 
 if HAVE_VERBS
 _usnic_files += prov/usnic/src/usdf_fake_ibv.c
